- cache

+++ app/src/main/java/es/jvbabi/vplanplus/data/repository/RoomRepositoryImpl.kt
@@ -151,7 +151,7 @@ class RoomRepositoryImpl(
         roomBookings
             .bookings
             .map { it.bookedBy }
-            .forEach { vppId -> vppIdRepository.getVppId(vppId, school, false) }
+            .forEach { vppId -> vppIdRepository.getVppId(vppId.toLong(), school, false) }
 
         val existingBookings = roomBookingDao.getAllRoomBookings().map { it.roomBooking.id }
 
@@ -179,7 +179,7 @@ class RoomRepositoryImpl(
         roomBookings.bookings
             .filter profileClass@{ booking -> booking.`class` in profileClasses }
             .filter notInPast@{ booking -> DateUtils.getDateTimeFromTimestamp(booking.end).isAfter(LocalDateTime.now()) }
-            .filter notBookedByCurrentUser@{ booking -> !(vppIdRepository.getVppId(booking.bookedBy, school, false)?.isActive() ?: false) }
+            .filter notBookedByCurrentUser@{ booking -> !(vppIdRepository.getVppId(booking.bookedBy.toLong(), school, false)?.isActive() ?: false) }
             .filter isNewInDatabase@{ booking -> !existingBookings.contains(booking.id) }
             .forEach sendNotification@{ booking ->
                 notificationRepository.sendNotification(
@@ -188,7 +188,7 @@ class RoomRepositoryImpl(
                     title = stringRepository.getString(R.string.notification_roomBookingTitle),
                     message = stringRepository.getString(
                         R.string.notification_roomBookingContent,
-                        vppIdRepository.getVppId(booking.bookedBy, school, false)?.name ?: stringRepository.getString(R.string.unknownVppId),
+                        vppIdRepository.getVppId(booking.bookedBy.toLong(), school, false)?.name ?: stringRepository.getString(R.string.unknownVppId),
                         booking.roomName,
                         DateUtils.getDateTimeFromTimestamp(booking.start).toLocalTime().format(
                             DateTimeFormatter.ofPattern("HH:mm")
